Refactor application setup to use WebApplication model
Significantly refactored `Program.cs` to utilize the new `WebApplication` model, consolidating service configuration and middleware setup directly within the `Main` method. Removed the `Startup` class, streamlining the application structure and enhancing clarity. Introduced new services for SaaS API configuration and fulfillment, along with improved logging setup.
